refactor: move ift query commands into ift.go
